requirements.txt
requests
python-dotenv
notion-client
beautifulsoup4
